+++ test_repo.py
@@ -3,7 +3,8 @@ import psycopg2
 from unittest.mock import MagicMock, patch
 from models import *
 from psycopg2 import pool
-from repository import Repository
+from repository.event_repository import EventRepository
+from repository.feedback_repository import FeedbackRepository
 
 class EventModelTest:
     def __init__(self, description=None, location=None, likes=None, event_date=None):
@@ -64,7 +65,7 @@ def test_events_get_all():
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchall.return_value = event_row
         cursor_mock.fetchone.return_value = [True]
-        repo = Repository()
+        repo = EventRepository()
         events = repo.get_events_all()
         assert events[0].title == event1.title
         assert events[1].description == event2.description
@@ -80,7 +81,7 @@ def test_events_get_by_id():
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchall.return_value = feedback_row
         cursor_mock.fetchone.return_value = event_row[0]
-        repo = Repository()
+        repo = EventRepository()
         event = repo.get_event_by_id(1)
         assert event.title == event1.title
         assert event.description == event1.description
@@ -95,7 +96,7 @@ def test_event_delete():
         cursor_mock = MagicMock()
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
-        repo = Repository()
+        repo = EventRepository()
         events = repo.delete_event(1)
         assert events == None
 
@@ -109,7 +110,7 @@ def test_update_event():
         cursor_mock = MagicMock()
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
-        repo = Repository()
+        repo = EventRepository()
         event = repo.update_event(event3.__dict__, 1)
         assert event == None
 
@@ -124,7 +125,7 @@ def test_add_event():
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchone.return_value = [3]
-        repo = Repository()
+        repo = EventRepository()
         event = repo.add_event(event3.__dict__)
         assert event.title == event3.title
         assert event.location == event3.location
@@ -143,7 +144,7 @@ def test_modify_event():
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchall.return_value = []
         cursor_mock.fetchone.return_value = event_row[0]
-        repo = Repository()
+        repo = EventRepository()
         event = repo.modify_event(eventPatch.__dict__, event1.id)
         assert event == None
         
@@ -158,7 +159,7 @@ def test_feedbacks_get_all():
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchall.return_value = feedback_row
         cursor_mock.fetchone.return_value = [True]
-        repo = Repository()
+        repo = FeedbackRepository()
         feedbacks = repo.get_feedbacks_all(1)
         assert feedbacks[0].content == feedback1.content
         assert feedbacks[1].created_at == feedback2.created_at
@@ -173,7 +174,7 @@ def test_feedback_get_by_id():
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchone.return_value = feedback_row[0]
-        repo = Repository()
+        repo = FeedbackRepository()
         feedback = repo.get_feedback_by_id(1)
         assert feedback.content == feedback1.content
         assert feedback.created_at == feedback1.created_at
@@ -187,7 +188,7 @@ def test_feedback_delete():
         cursor_mock = MagicMock()
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
-        repo = Repository()
+        repo = FeedbackRepository()
         feedback = repo.delete_feedback(1)
         assert feedback == None
 
@@ -200,7 +201,7 @@ def test_update_feedback():
         cursor_mock = MagicMock()
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
-        repo = Repository()
+        repo = FeedbackRepository()
         event = repo.update_feedback(feedback3.__dict__, 1)
         assert event == None
 
@@ -215,7 +216,7 @@ def test_add_feedback():
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchone.return_value = [3]
-        repo = Repository()
+        repo = FeedbackRepository()
         feedback = repo.add_feedback(feedback2.__dict__, 1)
         assert feedback.content == feedback2.content
         assert feedback.created_at == feedback2.created_at
@@ -232,7 +233,7 @@ def test_modify_feedback():
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchall.return_value = []
         cursor_mock.fetchone.return_value = feedback_row[0]
-        repo = Repository()
+        repo = FeedbackRepository()
         feedback = repo.modify_feedback(feedback3.__dict__, feedback1.id)
         assert feedback == None
         
@@ -246,7 +247,7 @@ def test_check_feedback_existence():
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchone.return_value = [False]
-        repo = Repository()
+        repo = FeedbackRepository()
         feedback = repo.check_feedback_existence(feedback1.id)
         assert feedback == False
 
@@ -260,7 +261,7 @@ def test_check_feedback_existence_by_event_id():
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchone.return_value = [False]
-        repo = Repository()
+        repo = FeedbackRepository()
         feedback = repo.check_feedback_existence_by_event_id(feedback1.id)
         assert feedback == False
         
@@ -274,7 +275,7 @@ def test_check_event_existence():
         p_mock.getconn.return_value = conn_mock
         conn_mock.cursor.return_value = cursor_mock
         cursor_mock.fetchone.return_value = [False]
-        repo = Repository()
+        repo = FeedbackRepository()
         feedback = repo.check_event_existence(feedback1.id)
         assert feedback == False
